+++ go.mod
@@ -7,7 +7,11 @@ toolchain go1.22.8
 require golang.org/x/tools v0.27.0
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golangci/plugin-module-register v0.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.10.0 // indirect
 	golang.org/x/mod v0.22.0 // indirect
 	golang.org/x/sync v0.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
